+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Add `BankQuerier` as trait bound to `CwEnv`.
 - Add `WasmCodeQuerier` as trait bound to `CwEnv`.
 - Changed the snapshot feature for better snapshot readability.
+- Added Readonly flag to the daemon state
+- Added min gas fee environment variable
 - Updated `cosmrs` to `0.15.0`
 - Updated tonic to `0.10.2`
 - Bumped MSRV to 1.72 because of dependency `cosmrs@0.15.0`